requirements.txt
streamlit
numpy
pandas
scikit-learn
scipy
joblib
liac-arff
